+++ slurp.c
@@ -15,7 +15,7 @@ slurp(char *file, size_t *np) {
         return NULL;
     }
     fseek (fp, 0, SEEK_END);
-    n = ftell (fp);
+    n = (size_t) ftell (fp);
     fseek (fp, 0, SEEK_SET);
     data = calloc(n, sizeof(char));
     if(fread (data, 1, n, fp) != n) {
